add all variant context in utility method
